+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Added rating filters feature for the store front
+
 ## [3.3.7] - 2021-12-22
 
 ### Fixed